# Storybook: Improve TypeScript performance for slow stories