--- a/packages/components/src/dropdown-menu/stories/index.story.tsx
+++ b/packages/components/src/dropdown-menu/stories/index.story.tsx
@@ -73,7 +73,6 @@ export const Default: StoryObj< typeof DropdownMenu > = {
 
 export const WithChildren: StoryObj< typeof DropdownMenu > = {
 	...Default,
-	controls: null,
 	// Adding custom source because Storybook is not able to show the contents of
 	// the `children` prop correctly in the code snippet.
 	parameters: {
@@ -100,7 +99,6 @@ export const WithChildren: StoryObj< typeof DropdownMenu > = {
 		},
 	},
 	args: {
-		...Default.args,
 		label: 'Select a direction.',
 		icon: more,
 		children: ( { onClose } ) => (