<!-- Thanks for contributing to Gutenberg! Please follow the Gutenberg Contributing Guidelines:
https://github.com/WordPress/gutenberg/blob/trunk/CONTRIBUTING.md -->

## What?

Optimizes some low-hanging bottlenecks in the TypeScript compilation performance for Storybook stories, as revealed by [`@typescript/analyze-trace`](https://github.com/microsoft/typescript-analyze-trace).

### Improvement in cold build time

```sh
# Before
npx tsc --build --force  145.83s user 10.52s system 132% cpu 1:58.13 total

# After
npx tsc --build --force  136.40s user 9.36s system 134% cpu 1:48.61 total
```

Honestly the build times are not very stable so I'm not sure how much we can read into it, but the analyzed output of `@typescript/analyze-trace` clearly shows improvements, so I think we can say that we're moving things in the correct direction.

## Why?

These particularly slow stories have TS performance issues in the components themselves, and we ultimately need to address the root cause. However, their Storybook stories can be optimized to "bypass" some of the particularly slow parts. Luckily, this mostly just involves converting the story format from [CSF2 to CSF3](https://storybook.js.org/docs/7/api/csf#upgrading-from-csf-2-to-csf-3), with a bit of attention given to minimizing unnecessary type checking work. And CSF3 is good for us to transition to anyway.

## Testing Instructions

Smoke test the affected stories to see that there are no regressions.

Code review should be easier with whitespace hidden.